+++ pyod/test/test_utility.py
@@ -21,8 +21,6 @@ import numpy as np
 # if pyod is installed, no need to use the following line
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
 
-from utils.data import generate_data
-from utils.data import evaluate_print
 from utils.utility import check_parameter
 from utils.utility import standardizer
 from utils.utility import get_label_n
@@ -31,97 +29,6 @@ from utils.utility import argmaxn
 from utils.utility import invert_order
 from utils.utility import check_detector
 from utils.utility import score_to_label
-
-
-class TestUtils(unittest.TestCase):
-    def setUp(self):
-        self.n_train = 1000
-        self.n_test = 500
-        self.contamination = 0.1
-
-        self.value_lists = [0.1, 0.3, 0.2, -2, 1.5, 0, 1, -1, -0.5, 11]
-
-    def test_data_generate(self):
-        X_train, y_train, X_test, y_test = \
-            generate_data(n_train=self.n_train,
-                          n_test=self.n_test,
-                          contamination=self.contamination)
-
-        assert_equal(y_train.shape[0], X_train.shape[0])
-        assert_equal(y_test.shape[0], X_test.shape[0])
-
-        assert_less_equal(self.n_train - X_train.shape[0], 1)
-        assert_equal(X_train.shape[1], 2)
-
-        assert_less_equal(self.n_test - X_test.shape[0], 1)
-        assert_equal(X_test.shape[1], 2)
-
-        out_perc = np.sum(y_train) / self.n_train
-        assert_allclose(self.contamination, out_perc, atol=0.01)
-
-        out_perc = np.sum(y_test) / self.n_test
-        assert_allclose(self.contamination, out_perc, atol=0.01)
-
-    def test_data_generate2(self):
-        X_train, y_train, X_test, y_test = \
-            generate_data(n_train=self.n_train,
-                          n_test=self.n_test,
-                          n_features=3,
-                          contamination=self.contamination)
-        assert_allclose(X_train.shape, (self.n_train, 3))
-        assert_allclose(X_test.shape, (self.n_test, 3))
-
-    def test_data_generate3(self):
-        X_train, y_train, X_test, y_test = \
-            generate_data(n_train=self.n_train,
-                          n_test=self.n_test,
-                          n_features=2,
-                          contamination=self.contamination,
-                          random_state=42)
-
-        X_train2, y_train2, X_test2, y_test2 = \
-            generate_data(n_train=self.n_train,
-                          n_test=self.n_test,
-                          n_features=2,
-                          contamination=self.contamination,
-                          random_state=42)
-
-        assert_allclose(X_train, X_train2)
-        assert_allclose(X_test, X_test2)
-        assert_allclose(y_train, y_train2)
-        assert_allclose(y_test, y_test2)
-
-    def test_argmaxn(self):
-        ind = argmaxn(self.value_lists, 3)
-        assert_equal(len(ind), 3)
-
-        ind = argmaxn(self.value_lists, 3)
-        assert_equal(np.sum(ind), np.sum([4, 6, 9]))
-
-        ind = argmaxn(self.value_lists, 3, order='asc')
-        assert_equal(np.sum(ind), np.sum([3, 7, 8]))
-
-        with assert_raises(ValueError):
-            argmaxn(self.value_lists, -1)
-        with assert_raises(ValueError):
-            argmaxn(self.value_lists, 20)
-
-    def test_evaluate_print(self):
-        X_train, y_train, X_test, y_test = generate_data(n_train=self.n_train,
-                                                         n_test=self.n_test,
-                                                         contamination=self.contamination)
-        evaluate_print('dummy', y_train, y_train * 0.1)
-
-    # TODO: remove temporarily for pleasing Travis integration
-    # def test_visualize(self):
-    #     X_train, y_train, X_test, y_test = generate_data(
-    #         n_train=self.n_train, n_test=self.n_test,
-    #         contamination=self.contamination)
-    #     visualize('dummy', X_train, y_train, X_test, y_test, y_train * 0.1,
-    #               y_test * 0.1, show_figure=False, save_figure=False)
-
-    def tearDown(self):
-        pass
 
 
 class TestParameters(unittest.TestCase):
@@ -270,6 +177,7 @@ class TestMetrics(unittest.TestCase):
         self.labels_short_ = [0.1, 0.2, 0.2, 0.8, 0.2, 0.5, 0.7, 0.9, 1]
         self.manual_labels = [0, 0, 0, 1, 0, 0, 1, 1, 1, 0]
         self.outliers_fraction = 0.3
+        self.value_lists = [0.1, 0.3, 0.2, -2, 1.5, 0, 1, -1, -0.5, 11]
 
     def test_precision_n_scores(self):
         assert_equal(precision_score(self.y, self.manual_labels),
@@ -295,6 +203,21 @@ class TestMetrics(unittest.TestCase):
 
         labels = score_to_label(manual_scores, outliers_fraction=0.3)
         assert_allclose(labels, [0, 0, 0, 0, 0, 1, 0, 1, 1, 0])
+
+    def test_argmaxn(self):
+        ind = argmaxn(self.value_lists, 3)
+        assert_equal(len(ind), 3)
+
+        ind = argmaxn(self.value_lists, 3)
+        assert_equal(np.sum(ind), np.sum([4, 6, 9]))
+
+        ind = argmaxn(self.value_lists, 3, order='asc')
+        assert_equal(np.sum(ind), np.sum([3, 7, 8]))
+
+        with assert_raises(ValueError):
+            argmaxn(self.value_lists, -1)
+        with assert_raises(ValueError):
+            argmaxn(self.value_lists, 20)
 
     def tearDown(self):
         pass
